// Desktop/BharatGodam_App-divyang/src/types/entities.ts
export interface MongoDocument {
  _id: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}
// ...
export interface Manager extends MongoDocument {
  name: string;
  email: string;
  password: string;
  role: string;
  isDocumentsVerified: boolean;
  isAddressVerified: boolean;
  isKycDone: boolean;
  lastLoginTime: string;
  isEmailVerified: boolean;
  isMobileVerified: boolean;
  isActive: boolean;
  isArchived: boolean;
  isVerified: boolean;
  isRejected: boolean;
  warehouse: string[];
}